Fix integer offset issue.
FITS doesn't strictly support unsigned ints, but instead it gets around this by setting an offset value to unsigned ints:
https://heasarc.gsfc.nasa.gov/docs/software/fitsio/c/c_user/node23.html
This is fine, except that if the user is reading the file via the PDS4 label byte locations (rather than a FITS file reader) they will get this offset value and may not be aware of the FITS offset issue.  Therefore, we are converting all ints to signed-ints to avoid this problem.  To ensure we don't have any loss of data or integer wrapping, unsigned-int values are converted to a higher bit signed-integer type.   For example a uint16 now becomes an int32.